# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"SaveUpdateRes": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"Name": {
						"type": "string"
					},
					"Cell": {
						"type": "string"
					},
					"Adress": {
						"type": "string"
					},
					"UserName": {
						"type": "string"
					},
					"Password": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"Name",
					"Cell",
					"Adress",
					"UserName",
					"Password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveReqCustomer": {
				"properties": {
					"Name": {
						"type": "string"
					},
					"Cell": {
						"type": "string"
					},
					"Adress": {
						"type": "string"
					},
					"UserName": {
						"type": "string"
					},
					"Password": {
						"type": "string"
					}
				},
				"required": [
					"Name",
					"Cell",
					"Adress",
					"UserName",
					"Password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateReqCustomer": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"Name": {
						"type": "string"
					},
					"Cell": {
						"type": "string"
					},
					"Adress": {
						"type": "string"
					},
					"UserName": {
						"type": "string"
					},
					"Password": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"Name",
					"Cell",
					"Adress",
					"UserName",
					"Password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LogInReq": {
				"properties": {
					"UserName": {
						"type": "string"
					},
					"Password": {
						"type": "string"
					}
				},
				"required": [
					"UserName",
					"Password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveUpdateResProduct": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"Name": {
						"type": "string"
					},
					"Price": {
						"type": "string"
					},
					"Discription": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"Name",
					"Price",
					"Discription"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveReqProduct": {
				"properties": {
					"Name": {
						"type": "string"
					},
					"Price": {
						"type": "string"
					},
					"Discription": {
						"type": "string"
					}
				},
				"required": [
					"Name",
					"Price",
					"Discription"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateReqProduct": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"Name": {
						"type": "string"
					},
					"Price": {
						"type": "string"
					},
					"Discription": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"Name",
					"Price",
					"Discription"
				],
				"type": "object",
				"additionalProperties": false
			},
			"_LeanDocument__LeanDocument_T__": {
				"properties": {},
				"type": "object"
			},
			"Pick__LeanDocument_T_.Exclude_keyof_LeanDocument_T_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested__": {
				"properties": {
					"_id": {
						"$ref": "#/components/schemas/_LeanDocument__LeanDocument_T__",
						"description": "This documents _id."
					},
					"__v": {
						"description": "This documents __v."
					},
					"id": {
						"description": "The string version of this documents _id."
					}
				},
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit__LeanDocument_this_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested_": {
				"$ref": "#/components/schemas/Pick__LeanDocument_T_.Exclude_keyof_LeanDocument_T_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"LeanDocument_this_": {
				"$ref": "#/components/schemas/Omit__LeanDocument_this_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested_"
			},
			"IProduct": {
				"$ref": "#/components/schemas/LeanDocument_this_"
			},
			"DeletReqProduct": {
				"properties": {
					"_id": {
						"type": "string"
					}
				},
				"required": [
					"_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetReqProduct": {
				"properties": {
					"_id": {
						"type": "string"
					}
				},
				"required": [
					"_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProductInterface": {
				"properties": {
					"product": {
						"type": "string"
					},
					"quantity": {
						"type": "number",
						"format": "double"
					},
					"price": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"product",
					"quantity",
					"price"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveUpdateResOrder": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"Product": {
						"items": {
							"$ref": "#/components/schemas/ProductInterface"
						},
						"type": "array"
					},
					"Customer": {
						"type": "string"
					},
					"Delivery": {
						"type": "string"
					},
					"Bill": {
						"type": "number",
						"format": "double"
					},
					"TotalBill": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"Product",
					"Customer",
					"Delivery",
					"Bill",
					"TotalBill"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveReqOrder": {
				"properties": {
					"Product": {
						"items": {
							"$ref": "#/components/schemas/ProductInterface"
						},
						"type": "array"
					},
					"Customer": {
						"type": "string"
					},
					"Delivery": {
						"type": "string"
					}
				},
				"required": [
					"Product",
					"Customer",
					"Delivery"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetReqOrder": {
				"properties": {
					"Id": {
						"type": "string"
					}
				},
				"required": [
					"Id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetcustomerOrder": {
				"properties": {
					"Id": {
						"type": "string"
					}
				},
				"required": [
					"Id"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "token",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "backend",
		"version": "1.0.0",
		"description": "this is backend project",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/customer/savecustomer": {
			"post": {
				"operationId": "SaveCustomer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateRes"
								}
							}
						}
					}
				},
				"tags": [
					"customer"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SaveReqCustomer"
							}
						}
					}
				}
			}
		},
		"/customer/updatcustomer": {
			"put": {
				"operationId": "UpdateCustomer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateRes"
								}
							}
						}
					}
				},
				"tags": [
					"customer"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateReqCustomer"
							}
						}
					}
				}
			}
		},
		"/customer/loginuser": {
			"post": {
				"operationId": "LoginUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateRes"
								}
							}
						}
					}
				},
				"tags": [
					"customer"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LogInReq"
							}
						}
					}
				}
			}
		},
		"/product/saveproduct": {
			"post": {
				"operationId": "Saveproduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResProduct"
								}
							}
						}
					}
				},
				"tags": [
					"product"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SaveReqProduct"
							}
						}
					}
				}
			}
		},
		"/product/updateproduct": {
			"put": {
				"operationId": "UpdateProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResProduct"
								}
							}
						}
					}
				},
				"tags": [
					"product"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateReqProduct"
							}
						}
					}
				}
			}
		},
		"/product/deleteproduct": {
			"delete": {
				"operationId": "Deletproduct",
				"responses": {
					"200": {
						"description": "Product is deleted",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IProduct"
										},
										{
											"properties": {
												"_id": {
													"type": "string"
												}
											},
											"required": [
												"_id"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"product"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeletReqProduct"
							}
						}
					}
				}
			}
		},
		"/product/getproduct": {
			"post": {
				"operationId": "Getproduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResProduct"
								}
							}
						}
					}
				},
				"tags": [
					"product"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GetReqProduct"
							}
						}
					}
				}
			}
		},
		"/product/getproductslist": {
			"get": {
				"operationId": "GetproductList",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SaveUpdateResProduct"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"product"
				],
				"security": [],
				"parameters": []
			}
		},
		"/order/saveorder": {
			"post": {
				"operationId": "SaveOrder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResOrder"
								}
							}
						}
					}
				},
				"tags": [
					"order"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SaveReqOrder"
							}
						}
					}
				}
			}
		},
		"/order/getorder": {
			"post": {
				"operationId": "GetOrder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResOrder"
								}
							}
						}
					}
				},
				"tags": [
					"order"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GetReqOrder"
							}
						}
					}
				}
			}
		},
		"/order/getAllorder": {
			"post": {
				"operationId": "GetOrderList",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SaveUpdateResOrder"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"order"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GetcustomerOrder"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}